Fix order transfer roomguest bug
Transferring the order to someone who was in a room as a simple (confirmed) guest, didn't delete the room_guest object, making him stuck in the room. Due to a second bug in the `getRoomGuestResponseFromRoomId` query, people without order were not listed as guests even if confirmed, making the users/admins unable to see the phantom guest and passing the sanity checks since the user (never returned as guest) is never checked if he has an order or not

Refer: `05ceaa05-bfb2-4c07-a4cc-aa9fc0cead04`

+++ application/src/main/java/net/furizon/backend/feature/room/RoomGeneralSanityCheck.java
@@ -29,21 +29,7 @@ import static net.furizon.backend.infrastructure.email.EmailVars.LINK;
 import static net.furizon.backend.infrastructure.email.EmailVars.OTHER_FURSONA_NAME;
 import static net.furizon.backend.infrastructure.email.EmailVars.ROOM_NAME;
 import static net.furizon.backend.infrastructure.email.EmailVars.SANITY_CHECK_REASON;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_NO_OWNER;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_HAS_DAILY_TICKET;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_HAS_NOT_BOUGHT_ROOM;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_HAS_NO_ORDER;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_IN_TOO_MANY_ROOMS;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_NOT_IN_ROOM;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_OWNER_ORDER_INVALID_STATUS;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_TOO_MANY_MEMBERS;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_USER_HAS_DAILY_TICKET;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_USER_HAS_NO_ORDER;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_USER_IN_TOO_MANY_ROOMS;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.SC_USER_ORDER_INVALID_ORDER_STATUS;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.TEMPLATE_SANITY_CHECK_DELETED;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.TEMPLATE_SANITY_CHECK_KICK_OWNER;
-import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.TEMPLATE_SANITY_CHECK_KICK_USER;
+import static net.furizon.backend.infrastructure.rooms.RoomEmailTexts.*;
 
 @Slf4j
 @Component
@@ -184,6 +170,18 @@ public class RoomGeneralSanityCheck {
                 sanityCheckLogAndStoreErrors(detectedErrors, "[ROOM SANITY CHECKS] "
                         + "Room {} has too many members ({} > {}). Deleting the room", roomId, guestNo, capacity);
                 mails.addAll(prepareSanityCheckMailRoomDeleted(roomId, roomName, SC_TOO_MANY_MEMBERS));
+                roomLogic.deleteRoom(roomId);
+                return;
+            }
+
+            int confirmedGuestsNo = roomFinder.getRoomGuestsFromRoomId(roomId, true).size();
+            if (confirmedGuestsNo != guestNo) {
+                //delete room, number of confirmed guests object mismatch the getRoomGuestResponseFromRoomId size!
+                sanityCheckLogAndStoreErrors(detectedErrors, "[ROOM SANITY CHECKS] "
+                        + "Room {} number of confirmed guests objects ({}) mismatches "
+                        + "the getRoomGuestResponseFromRoomId's size ({}). Deleting the room",
+                        roomId, confirmedGuestsNo, guestNo);
+                mails.addAll(prepareSanityCheckMailRoomDeleted(roomId, roomName, SC_ROOM_GUESTS_MISMATCH));
                 roomLogic.deleteRoom(roomId);
                 return;
             }
